Register and unregister function table callbacks at the code chunk level
Windows x64 only: Stack unwinding information for SEH is provided to
Windows via a callback function. Currently we install this callback for
every JITted function individually, at the same time as we calculate the
unwind information. Aside from this being inefficient, we also never
unregister the callback - meaning that after a domain reload, Windows
may ask us to provide unwind information for code that does not belong
to us, but just happens to have been loaded into addresses we used to
own (case 800613).

This commit changes the registration to be done at the code-chunk level
instead, such that we register the callback once per chunk, and also
unregister it again when the chunk is freed.

The unwind information itself is still calculated exactly as before and
stored in the same location (following the generated code itself). The
RUNTIME_FUNCTION structure is also stored in the same place, but because
it stores addresses relative to the one in the function table (which is
now the start of the entire code chunk, rather than the start of the
specific function being unwound), the members are updated accordingly.
In principle this could be done only when the function is first JITted,
but looking up the page to use would add a cost to the JIT of every
function, while this way we only add a much smaller cost to the specific
functions being unwound.