Removing invalid null check of hostname
The check if `*hostname` is set causes problems when `hostname`
points to an empty string. It will be then be false and cause
the following call to `mono_get_address_info` not to be made
resulting in `info` will not be filled with proper value but
point to null. Also, there is no point of checking if hostname
is set here since it's content is already used earlier in the
method.