[runtime] Remove PreserveAll Calling Convention on linux
On linux, this calling convention will damage the stack layout.
This does not cause a crash, as the damage is undone before returning
from the function. When libunwind or another unwinder tries to walk the
stack though, they'll see a frame with a reference to an ip that is not
in mapped memory. This causes stack walking functions to abort early.
In our case, this was the c++ exception handling walker that gave up
on finding a landing pad. Disabling it has a performance cost, but
is necessary for correctness.