Fixes socket timeout on non-blocking Connect()
Previously, a non-blocking Connect() call using the IPAddress[] overload
would block forever. This does not match either the mono
Connect(IPEndPoint) overload or the .NET Core and Windows .NET Framework
implementations, which throw the socket exception immediately on a
Connect() error. This fixes that error, and allows for a timeout on the
Connect() method.

This change is released under the MIT license.