[SRE] Make the TypeBuilderInst types user types, handling them in the BCL instead of the runtime.
This works by assigning pseudo tokens for them in ModuleBuilder.GetToken (), so they no longer need to be handled in the GetToken () icall.
Real tokens are assigned after the dynamic types they depend on have been created, by creating the corresponding finished reflection object
(i.e. a MonoMethod for a MethodOnTypeBuilderInst etc.) and obtaining its token.
In Save mode, this is done in ModuleBuilder.Save ().
In Run mode, this is done on-demand when the JIT tries to resolve a token into its corresponding metadata object, by the runtime calling
the RuntimeResolve () method on the builder object.