Save and restore r8:r11 on AMD64
We found that the new C# compiler inserts sequence points in a few
places where the old compiler did not. These additional sequence points
cause the context to be switched more often. Since the third integer
argument for a function call is passed in r8 in the Windows AMD64 ABI,
we found that not preserving r8 across a sequence point could corrupt
the value of that third argument when the debugger is attached.

So this issue has probably existed for some time, but we've just
noticed is now with the new C# compiler and the new sequence points.

The fix here is the preserve r8 (as well as r9, r10, and r11). We may
not need to preserve r10 and r11, since they are not used for function
arguments. However, there seems to be little cost to preserving them as
well, so I've kept them for completeness.

This corrects case 825988, case 828073, and case 816067 in Unity.