[sgen] Remove old code dealing with OOM
Back in the day, we used to pin objects if we fail to move them. In the evacuation case it made sense because it was stopping futher evacuation (we no longer evacuate into new blocks though). In the promotion case it was signaling afterwards the triggering of a major in order to free space for new promoted objects. This is unlikely to work since we might need to allocate gray queue sections or other data structures anyways as part of the collection. Avoid complicating the code if we don't get any guarantees.

The way we should deal with OOM is by trying to enforce the soft heap limit, which should always be set, either by the user or by us. If we still fail to keep the heap from growing even in the soft-heap-limit exceeded case (when we have small allowances and do repeated majors), there is not much we can do anyway, so handling failed allocation won't help with anything.